html/browsers/browsing-the-web/navigating-across-documents/002.html

<!doctype html>
<!--/
specid:
  - id: browsing-the-web.html#navigating-across-documents
    step: [6]

description: |
  1. The iframe src is set to 002-1.html (with a response that will be
      delayed by 1s)
  2. Before that navigation is complete (~500ms later), the iframe src will be
      set to 002-2.html, which **must** cancel the "not yet mature" navigation.

/-->
<title>Multiple simultaneous navigations</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="about:blank"></iframe>
<script>
var test = async_test();
var iframe = document.querySelector("iframe");

window.onload = test.step_func(function() {
  setTimeout(test.step_func(function() {
    iframe.src = "002-1.html?pipe=trickle(d1)";
    setTimeout(test.step_func(function() {
      iframe.src = "002-2.html"
    }), 500);
  }), 100);
});

window.onmessage = test.step_func(function(e) {
  assert_equals(e.data, "002-2");
  test.done();
});

</script>
